pyproject.toml
[tool.poetry]
name = "desafio-techbiz"
version = "0.1.0"
description = ""
authors = ["Aldo Nunes"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langchain = "^0.2.14"
langchain-openai = "^0.1.22"
langchain-community = "^0.2.12"
unstructured = {extras = ["csv"], version = "^0.15.7"}
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
